requirements.txt
asgiref==3.7.2
black==23.9.1
cfgv==3.4.0
click==8.1.7
distlib==0.3.7
Django==4.2
django-environ==0.10.0
django-filter==23.3
djangorestframework==3.14.0
filelock==3.12.4
identify==2.5.30
importlib-metadata==6.8.0
isort==5.12.0
Markdown==3.4.4
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.11.2
petl==1.7.12
platformdirs==3.11.0
pre-commit==3.4.0
psycopg2-binary==2.9.6
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
virtualenv==20.24.5
zipp==3.17.0
